moved libp2pConnectionMonitor.go and test file into a new subpackage
