+++ p2p/libp2p/connectionMonitor/libp2pConnectionMonitor.go
@@ -1,4 +1,4 @@
-package libp2p
+package connectionMonitor
 
 import (
 	"math"
